Only cleanup ocp/tf-master if its there
This failed if the directory didn't exist

+++ masters_cleanup.sh
@@ -2,8 +2,10 @@
 
 set -eux
 
-pushd ocp/tf-master
-terraform init  # in case plugin has changed
-terraform destroy --auto-approve
-popd
-rm -rf ocp/tf-master
+if [ -d ocp/tf-master ] ; then
+    pushd ocp/tf-master
+    terraform init  # in case plugin has changed
+    terraform destroy --auto-approve
+    popd
+    rm -rf ocp/tf-master
+fi
